+++ src/Line.php
@@ -6,52 +6,24 @@ namespace Location;
 
 use Location\Bearing\BearingInterface;
 use Location\Distance\DistanceInterface;
+use Location\Intersection\Intersection;
 use Location\Utility\Cartesian;
 use RuntimeException;
 
-/**
- * Line Implementation
- *
- * @author Marcus Jaschen <mjaschen@gmail.com>
- */
-class Line implements GeometryInterface
+class Line implements GeometryLinesInterface
 {
     use GetBoundsTrait;
 
-    /**
-     * @var Coordinate
-     */
-    protected $point1;
+    final public const ORIENTATION_COLLINEAR = 0;
+    final public const ORIENTATION_CLOCKWISE = 1;
+    final public const ORIENTATION_ANTI_CLOCKWISE = 2;
 
-    /**
-     * @var Coordinate
-     */
-    protected $point2;
-
-    /**
-     * @param Coordinate $point1
-     * @param Coordinate $point2
-     */
-    public function __construct(Coordinate $point1, Coordinate $point2)
+    public function __construct(public readonly Coordinate $point1, public readonly Coordinate $point2)
     {
-        $this->point1 = $point1;
-        $this->point2 = $point2;
     }
 
     /**
-     * @param Coordinate $point1
-     *
-     * @return void
-     *
-     * @deprecated
-     */
-    public function setPoint1(Coordinate $point1)
-    {
-        $this->point1 = $point1;
-    }
-
-    /**
-     * @return Coordinate
+     * @deprecated Use property instead
      */
     public function getPoint1(): Coordinate
     {
@@ -59,19 +31,7 @@ class Line implements GeometryInterface
     }
 
     /**
-     * @param Coordinate $point2
-     *
-     * @return void
-     *
-     * @deprecated
-     */
-    public function setPoint2(Coordinate $point2)
-    {
-        $this->point2 = $point2;
-    }
-
-    /**
-     * @return Coordinate
+     * @deprecated Use property instead
      */
     public function getPoint2(): Coordinate
     {
@@ -89,33 +49,31 @@ class Line implements GeometryInterface
     }
 
     /**
+     * Returns an array containing the line segment.
+     *
+     * @return array<Line>
+     */
+    public function getSegments(): array
+    {
+        return [$this];
+    }
+
+    /**
      * Calculates the length of the line (distance between the two
-     * coordinates).
+     * coordinates in meters).
      *
      * @param DistanceInterface $calculator instance of distance calculation class
-     *
-     * @return float
      */
     public function getLength(DistanceInterface $calculator): float
     {
         return $calculator->getDistance($this->point1, $this->point2);
     }
 
-    /**
-     * @param BearingInterface $bearingCalculator
-     *
-     * @return float
-     */
     public function getBearing(BearingInterface $bearingCalculator): float
     {
         return $bearingCalculator->calculateBearing($this->point1, $this->point2);
     }
 
-    /**
-     * @param BearingInterface $bearingCalculator
-     *
-     * @return float
-     */
     public function getFinalBearing(BearingInterface $bearingCalculator): float
     {
         return $bearingCalculator->calculateFinalBearing($this->point1, $this->point2);
@@ -123,8 +81,6 @@ class Line implements GeometryInterface
 
     /**
      * Create a new instance with reversed point order, i. e. reversed direction.
-     *
-     * @return Line
      */
     public function getReverse(): Line
     {
@@ -135,8 +91,6 @@ class Line implements GeometryInterface
      * Get the midpoint of a Line segment
      *
      * @see http://www.movable-type.co.uk/scripts/latlong.html#midpoint
-     *
-     * @return Coordinate
      */
     public function getMidpoint(): Coordinate
     {
@@ -150,8 +104,8 @@ class Line implements GeometryInterface
         $B = new Cartesian(cos($lat2) * cos($deltaLng), cos($lat2) * sin($deltaLng), sin($lat2));
         $C = $A->add($B);
 
-        $latMid = atan2($C->getZ(), sqrt($C->getX() ** 2 + $C->getY() ** 2));
-        $lngMid = $lng1 + atan2($C->getY(), $C->getX());
+        $latMid = atan2($C->z, sqrt($C->x ** 2 + $C->y ** 2));
+        $lngMid = $lng1 + atan2($C->y, $C->x);
 
         return new Coordinate(rad2deg($latMid), rad2deg($lngMid));
     }
@@ -164,8 +118,6 @@ class Line implements GeometryInterface
      * @see http://www.edwilliams.org/avform.htm#Intermediate
      *
      * @param float $fraction 0.0 ... 1.0 (smaller or larger values work too)
-     *
-     * @return Coordinate
      *
      * @throws RuntimeException
      */
@@ -199,5 +151,67 @@ class Line implements GeometryInterface
         $lng = atan2($y, $x);
 
         return new Coordinate(rad2deg($lat), rad2deg($lng));
+    }
+
+    /**
+     * Compares the location of a given coordinate to this line returning
+     * its orientation as:
+     *
+     * - 0 if the coordinate is collinear to this line segment
+     * - 1 if the coordinate's orientation is clockwise to this line segment
+     * - 2 if the coordinate's orientation is anti-clockwise to this line segment
+     */
+    public function getOrientation(Coordinate $coordinate): int
+    {
+        $crossproduct1 = ($this->point2->getLat() - $this->point1->getLat())
+                         * ($coordinate->getLng() - $this->point2->getLng());
+        $crossproduct2 = ($this->point2->getLng() - $this->point1->getLng())
+                         * ($coordinate->getLat() - $this->point2->getLat());
+        $delta = $crossproduct1 - $crossproduct2;
+
+        if ($delta > 0) {
+            return self::ORIENTATION_CLOCKWISE;
+        }
+
+        if ($delta < 0) {
+            return self::ORIENTATION_ANTI_CLOCKWISE;
+        }
+
+        return self::ORIENTATION_COLLINEAR;
+    }
+
+    /**
+     * Two lines intersect if:
+     *
+     * 1. the points of the given line are oriented into opposite directions
+     * 2. the points of this line are oriented into opposite directions
+     * 3. the points are collinear and the two line segments are overlapping
+     */
+    public function intersectsLine(Line $line): bool
+    {
+        $orientation = [];
+        $orientation[11] = $this->getOrientation($line->point1);
+        $orientation[12] = $this->getOrientation($line->point2);
+        $orientation[21] = $line->getOrientation($this->point1);
+        $orientation[22] = $line->getOrientation($this->point2);
+
+        // the lines cross
+        if (
+            $orientation[11] !== $orientation[12]
+            && $orientation[21] !== $orientation[22]
+        ) {
+            return true;
+        }
+
+        // the lines are collinear or touch
+        if (
+            in_array(self::ORIENTATION_COLLINEAR, $orientation, true)
+            && (new Intersection())->intersects($this, $line, false)
+        ) {
+            return true;
+        }
+
+        // the lines do not overlap
+        return false;
     }
 }