+++ src/Polygon.php
@@ -7,24 +7,16 @@ namespace Location;
 use Location\Distance\DistanceInterface;
 use Location\Formatter\Polygon\FormatterInterface;
 
-/**
- * Polygon Implementation
- *
- * @author Paul Vidal <paul.vidal.lujan@gmail.com>
- * @author Marcus Jaschen <mjaschen@gmail.com>
- */
-class Polygon implements GeometryInterface
+class Polygon implements GeometryLinesInterface
 {
     use GetBoundsTrait;
 
     /**
      * @var array<Coordinate>
      */
-    protected $points = [];
+    protected array $points = [];
 
     /**
-     * @param Coordinate $point
-     *
      * @return void
      */
     public function addPoint(Coordinate $point): void
@@ -37,9 +29,7 @@ class Polygon implements GeometryInterface
      */
     public function addPoints(array $points): void
     {
-        foreach ($points as $point) {
-            $this->addPoint($point);
-        }
+        $this->points = [...$this->points, ...$points];
     }
 
     /**
@@ -60,7 +50,6 @@ class Polygon implements GeometryInterface
         $lats = [];
 
         foreach ($this->points as $point) {
-            /** @var Coordinate $point */
             $lats[] = $point->getLat();
         }
 
@@ -77,26 +66,17 @@ class Polygon implements GeometryInterface
         $lngs = [];
 
         foreach ($this->points as $point) {
-            /** @var Coordinate $point */
             $lngs[] = $point->getLng();
         }
 
         return $lngs;
     }
 
-    /**
-     * @return int
-     */
     public function getNumberOfPoints(): int
     {
         return count($this->points);
     }
 
-    /**
-     * @param FormatterInterface $formatter
-     *
-     * @return string
-     */
     public function format(FormatterInterface $formatter): string
     {
         return $formatter->format($this);
@@ -120,7 +100,7 @@ class Polygon implements GeometryInterface
 
         // to close the polygon we have to add the final segment between
         // the last point and the first point
-        $segments[] = new Line($this->points[$i - 1], $this->points[0]);
+        $segments[] = new Line($this->points[$length - 1], $this->points[0]);
 
         return $segments;
     }
@@ -133,10 +113,6 @@ class Polygon implements GeometryInterface
      *
      * - it's not detected when a line between two points is outside the polygon
      * - @see contains() for more restrictions
-     *
-     * @param GeometryInterface $geometry
-     *
-     * @return bool
      */
     public function containsGeometry(GeometryInterface $geometry): bool
     {
@@ -159,16 +135,12 @@ class Polygon implements GeometryInterface
      * For special cases this calculation leads to wrong results:
      *
      * - if the polygons spans over the longitude boundaries at 180/-180 degrees
-     *
-     * @param Coordinate $point
-     *
-     * @return bool
      */
     public function contains(Coordinate $point): bool
     {
         $numberOfPoints = $this->getNumberOfPoints();
-        $polygonLats    = $this->getLats();
-        $polygonLngs    = $this->getLngs();
+        $polygonLats = $this->getLats();
+        $polygonLngs = $this->getLngs();
 
         $polygonContainsPoint = false;
 
@@ -179,7 +151,7 @@ class Polygon implements GeometryInterface
                     / ($polygonLngs[$altNode] - $polygonLngs[$node]) + $polygonLats[$node]);
 
             if ($condition) {
-                $polygonContainsPoint = ! $polygonContainsPoint;
+                $polygonContainsPoint = !$polygonContainsPoint;
             }
         }
 
@@ -190,8 +162,6 @@ class Polygon implements GeometryInterface
      * Calculates the polygon perimeter.
      *
      * @param DistanceInterface $calculator instance of distance calculation class
-     *
-     * @return float
      */
     public function getPerimeter(DistanceInterface $calculator): float
     {
@@ -214,8 +184,6 @@ class Polygon implements GeometryInterface
      * This algorithm gives inaccurate results as it ignores
      * ellipsoid parameters other than to arithmetic mean radius.
      * The error should be < 1 % for small areas.
-     *
-     * @return float
      */
     public function getArea(): float
     {
@@ -226,12 +194,12 @@ class Polygon implements GeometryInterface
         }
 
         $referencePoint = $this->points[0];
-        $radius         = $referencePoint->getEllipsoid()->getArithmeticMeanRadius();
-        $segments       = $this->getSegments();
+        $radius = $referencePoint->getEllipsoid()->getArithmeticMeanRadius();
+        $segments = $this->getSegments();
 
         foreach ($segments as $segment) {
-            $point1 = $segment->getPoint1();
-            $point2 = $segment->getPoint2();
+            $point1 = $segment->point1;
+            $point2 = $segment->point2;
 
             $x1 = deg2rad($point1->getLng() - $referencePoint->getLng()) * cos(deg2rad($point1->getLat()));
             $y1 = deg2rad($point1->getLat() - $referencePoint->getLat());
@@ -250,8 +218,6 @@ class Polygon implements GeometryInterface
     /**
      * Create a new polygon with reversed order of points, i. e. reversed
      * polygon direction.
-     *
-     * @return Polygon
      */
     public function getReverse(): Polygon
     {