+++ src/Utility/PointToLineDistance.php
@@ -6,53 +6,64 @@ namespace Location\Utility;
 
 use Location\Coordinate;
 use Location\Distance\DistanceInterface;
+use Location\Exception\NotConvergingException;
 use Location\Line;
 
-/**
- * Calculate the distance between a Line and a Point.
- *
- * @author Marcus Jaschen <mjaschen@gmail.com>
- */
 class PointToLineDistance
 {
-    /**
-     * @var DistanceInterface
-     */
-    private $distanceCalculator;
-
-    public function __construct(DistanceInterface $distanceCalculator)
-    {
-        $this->distanceCalculator = $distanceCalculator;
+    public function __construct(
+        private readonly DistanceInterface $distanceCalculator,
+        private readonly float $epsilon = 0.001
+    ) {
     }
 
+    /**
+     * @psalm-suppress InvalidReturnType
+     * @throws NotConvergingException
+     */
     public function getDistance(Coordinate $point, Line $line): float
     {
-        if ($line->getPoint1()->hasSameLocation($line->getPoint2())) {
-            return $this->distanceCalculator->getDistance($point, $line->getPoint1());
+        if ($line->point1->hasSameLocation($line->point2, $this->epsilon)) {
+            return $this->distanceCalculator->getDistance($point, $line->point1);
+        }
+        if ($point->hasSameLocation($line->point1, $this->epsilon)) {
+            return 0.0;
+        }
+        if ($point->hasSameLocation($line->point2, $this->epsilon)) {
+            return 0.0;
+        }
+        if ($point->hasSameLocation($line->getMidpoint(), $this->epsilon)) {
+            return 0.0;
         }
 
-        $pLat = deg2rad($point->getLat());
-        $pLng = deg2rad($point->getLng());
+        $iterationsCounter = 0;
+        $iterationLine = clone $line;
 
-        $l1Lat = deg2rad($line->getPoint1()->getLat());
-        $l1Lng = deg2rad($line->getPoint1()->getLng());
-        $l2Lat = deg2rad($line->getPoint2()->getLat());
-        $l2Lng = deg2rad($line->getPoint2()->getLng());
+        do {
+            $linePoint1 = $iterationLine->point1;
+            $linePoint2 = $iterationLine->point2;
+            $lineMidPoint = $iterationLine->getMidpoint();
 
-        $deltal2l1Lat = $l2Lat - $l1Lat;
-        $deltal2l1Lng = $l2Lng - $l1Lng;
+            $distancePointToLinePoint1 = $point->getDistance($linePoint1, $this->distanceCalculator);
+            $distancePointToLinePoint2 = $point->getDistance($linePoint2, $this->distanceCalculator);
 
-        $u = (($pLat - $l1Lat) * $deltal2l1Lat + ($pLng - $l1Lng) * $deltal2l1Lng) /
-            ($deltal2l1Lat ** 2 + $deltal2l1Lng ** 2);
+            if ($distancePointToLinePoint1 <= $distancePointToLinePoint2) {
+                $iterationLine = new Line($linePoint1, $lineMidPoint);
+            } else {
+                $iterationLine = new Line($lineMidPoint, $linePoint2);
+            }
 
-        if ($u <= 0) {
-            return $this->distanceCalculator->getDistance($point, $line->getPoint1());
-        }
+            if (abs($distancePointToLinePoint1 - $distancePointToLinePoint2) < $this->epsilon) {
+                return $point->getDistance($iterationLine->getMidpoint(), $this->distanceCalculator);
+            }
 
-        if ($u >= 1) {
-            return $this->distanceCalculator->getDistance($point, $line->getPoint2());
-        }
-
-        return (new PerpendicularDistance())->getPerpendicularDistance($point, $line);
+            $iterationsCounter++;
+            if ($iterationsCounter > 100) {
+                throw new NotConvergingException(
+                    'Calculation of Point to Minor Arc did not converge after 100 iterations.',
+                    6391878367
+                );
+            }
+        } while (true);
     }
 }